create LaunchedByDetail to consolidate shared code between detail & list
