+++ awx/ui_next/src/components/JobList/JobListItem.jsx
@@ -8,40 +8,11 @@ import { RocketIcon } from '@patternfly/react-icons';
 import { ActionsTd, ActionItem } from '../PaginatedTable';
 import LaunchButton from '../LaunchButton';
 import StatusLabel from '../StatusLabel';
-import { DetailList, Detail } from '../DetailList';
+import { DetailList, Detail, LaunchedByDetail } from '../DetailList';
 import ChipGroup from '../ChipGroup';
 import CredentialChip from '../CredentialChip';
 import { formatDateString } from '../../util/dates';
 import { JOB_TYPE_URL_SEGMENTS } from '../../constants';
-
-const getLaunchedByDetails = ({ summary_fields = {}, related = {} }) => {
-  const {
-    created_by: createdBy,
-    job_template: jobTemplate,
-    schedule,
-  } = summary_fields;
-  const { schedule: relatedSchedule } = related;
-
-  if (!createdBy && !schedule) {
-    return {};
-  }
-
-  let link;
-  let value;
-
-  if (createdBy) {
-    link = `/users/${createdBy.id}`;
-    value = createdBy.username;
-  } else if (relatedSchedule && jobTemplate) {
-    link = `/templates/job_template/${jobTemplate.id}/schedules/${schedule.id}`;
-    value = schedule.name;
-  } else {
-    link = null;
-    value = schedule.name;
-  }
-
-  return { link, value };
-};
 
 function JobListItem({
   i18n,
@@ -63,8 +34,6 @@ function JobListItem({
     workflow_job: i18n._(t`Workflow Job`),
   };
 
-  const { value: launchedByValue, link: launchedByLink } =
-    getLaunchedByDetails(job) || {};
   const { credentials, inventory, labels } = job.summary_fields;
 
   return (
@@ -141,16 +110,7 @@ function JobListItem({
                 label={i18n._(t`Finished`)}
                 value={formatDateString(job.started)}
               />
-              <Detail
-                label={i18n._(t`Launched By`)}
-                value={
-                  launchedByLink ? (
-                    <Link to={`${launchedByLink}`}>{launchedByValue}</Link>
-                  ) : (
-                    launchedByValue
-                  )
-                }
-              />
+              <LaunchedByDetail job={job} i18n={i18n} />
               {credentials && credentials.length > 0 && (
                 <Detail
                   fullWidth
